// web/src/api/generated/models/SharedPhoto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PhotoParentType } from './PhotoParentType';
import {
    PhotoParentTypeFromJSON,
    PhotoParentTypeFromJSONTyped,
    PhotoParentTypeToJSON,
    PhotoParentTypeToJSONTyped,
} from './PhotoParentType';
import type { AttachmentSummary } from './AttachmentSummary';
import {
    AttachmentSummaryFromJSON,
    AttachmentSummaryFromJSONTyped,
    AttachmentSummaryToJSON,
    AttachmentSummaryToJSONTyped,
} from './AttachmentSummary';

/**
 * 
 * @export
 * @interface SharedPhoto
 */
export interface SharedPhoto {
    /**
     * 
     * @type {AttachmentSummary}
     * @memberof SharedPhoto
     */
    attachment: AttachmentSummary;
    /**
     * 
     * @type {Date}
     * @memberof SharedPhoto
     */
    effectiveDate: Date;
    /**
     * 
     * @type {string}
     * @memberof SharedPhoto
     */
    parentId: string;
    /**
     * 
     * @type {PhotoParentType}
     * @memberof SharedPhoto
     */
    parentType: PhotoParentType;
}



/**
 * Check if a given object implements the SharedPhoto interface.
 */
export function instanceOfSharedPhoto(value: object): value is SharedPhoto {
    if (!('attachment' in value) || value['attachment'] === undefined) return false;
    if (!('effectiveDate' in value) || value['effectiveDate'] === undefined) return false;
    if (!('parentId' in value) || value['parentId'] === undefined) return false;
    if (!('parentType' in value) || value['parentType'] === undefined) return false;
    return true;
}

export function SharedPhotoFromJSON(json: any): SharedPhoto {
    return SharedPhotoFromJSONTyped(json, false);
}

export function SharedPhotoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SharedPhoto {
    if (json == null) {
        return json;
    }
    return {
        
        'attachment': AttachmentSummaryFromJSON(json['attachment']),
        'effectiveDate': (new Date(json['effectiveDate'])),
        'parentId': json['parentId'],
        'parentType': PhotoParentTypeFromJSON(json['parentType']),
    };
}

export function SharedPhotoToJSON(json: any): SharedPhoto {
    return SharedPhotoToJSONTyped(json, false);
}

export function SharedPhotoToJSONTyped(value?: SharedPhoto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'attachment': AttachmentSummaryToJSON(value['attachment']),
        'effectiveDate': value['effectiveDate'].toISOString().substring(0,10),
        'parentId': value['parentId'],
        'parentType': PhotoParentTypeToJSON(value['parentType']),
    };
}